requirements.txt
streamlit
pinecone-client[grpc]
langchain
openai
tiktoken